Fix Beatport purchased syncing;
Beatport has replaced the My Downloads page with a My Library page. The
new page fetches the tracks using an XHR and thus the logic in the sync
can be simplified. The format of JSON returned from the API is however
different than that of the Playables in My Downloads and thus requires
a different transformation.