Fix total duration fetching in preview player;
The total duration in the state refers to the total duration of the
preview clip whil the currentTrack duration refers to the duration of
the whole track (i.e. not the preview). Thus when there is no
pre-rendered waveform for the track and the waveform is created from the
preview, the duration needs to match the duration of the preview in
order for the preview waveform to be displayed correctly.