Fetch preview file url from backend only if it is null;
Now that the preview url is nullable in the database, the frontend can
determine whether it should get the preview file url from the backend
or use the one in the response. This way there is no unnecessary round
trip for stores whose preview urls can be cached.

This commit also fixes the showing of waveforms in the UI.