Do not try to fetch tracks for Bandcamp artists without url;
If the artists are not added to the db explicitly i.e. using the form
on the settings page, their url will not get set (difficult to extract
it from the details). Fetching the tracks for the artists is done by
url and thus is not possible if the url is not set. The Bandcamp artist
needs to also be excluded from the follows on settings page in such a
case to correctly communicate which artists' tracks will get fetched.