HACKFIX: Prevent player refresh when navigating back from settings;
When the "Tracks" -menu button links to the app root, the browser will be redirected
to /new when the buttion is clicked. This will trigger a reload and cause delay and
flickering in the transition. This commit changes the button to link to /new directly. This
however will have the issue that the state of the app will not match the url. Also it
seems that changing the tabs in settings causes the tab focus in tracks to disappear.