BUGFIX: Fix settings page loading by url;
Previously the /settings path was handled separately from the /:path.
This however meant that if the site was opened (or refreshed) with
the /settings path, the UI would fail to load. This change joins the
/:path and /settings routes and seems to fix that and also the issue
with displaying the incorrect tab for the tracks view when changing
pages in the settings.