Fix / unify score materialized views;
Hard to tell what happened here, but I guess the materialized views were
not updated when the score calculation logic was updated. Anyhow,
gapping the day difference to 60 will probably be difficult to explain
to the users and it is probably best to just have a small weight to
limit the score. On top of that the calculation was just plain wrong...