Try to fix Spotify playlist track response logging;
Currently the logging shows that the items list in the response is transformed to
an object. This might either be caused by Winston or some other peculiarity in
the handling of the response. This commit changes the logging to log an object
instead of the array in order to see whether the list still shows as an object
and not as an array as it should.