Try to fix Spotify playlist details fetching by providing market option;
For some reason it seems Spotify API no longer returns the preview_urls for tracks like it used to.
If you provide the market parameter in the request however, the preview urls are in the response.
This commit hardcodes the requests to use the US market with an assumption that most of the tracks
should be available in that market. Some online discussion points to using different authentication
flows or requesting the details on per track basis: https://github.com/spotify/web-api/issues/148.