Try to fix the Bandcamp track transform;
The Chrome extension probably calls the transform with release details for multiple releases
and thus the release details need to be wrapped in an array for the transform to work. Also
the regex seemed to always return 6 matches and thus it's necessary to also check that the
fifth element is not undefined before calling split for it.