WIP: Infinite scroll;
The infinite scroll is a bit difficult as the query searches for only tracks that
have not been heard yet. If the user listens to some tracks and then tries to get
the next tracks with a offset (i.e. pagination), some tracks will be skipped.

Also this functionality does not make much sense as the tracks at the top of the
list should be the most important and if those do not interest the user, the
sorting algorithm is not good enough and should be improved.

What would be useful is to fetch the tracks in smaller batches and limit the
amount of tracks rendered in the DOM, although in this case also you'd need
to take into account that the user might have listened to some tracks already.