Remove sorting parameter from /me/tracks;
The sorting makes little sense as it only applies to the new tracks
and those should anyway be sorted according to the starred status and
the sorting score.

This commit also gives a higher precedence for tracks where multiple
artists are starred.

A composableSql tagged template is provided for composition of sql``
string. With sql-template-strings it is only possible to join queries
with query.append, which makes the queries especially messy if one needs
to reuse a part of the query or define fields according to variables. In
order to reuse a query, you can first define it:

const order = sql`order by 1`

and then include it in another query:

const query = composableSql`select * from tab ${order}`

If you need to pass user input to the query it is better done like so:

const query = composableSql`select * from tab where col = ${sql`${val}`}`

If you do not use the sql template string here, the input will be passed
to the query as is (not in the values array) and thus will not be
escaped and may create an injection issue.