Postpone cart track fetch;
Fetching the tracks for all carts takes quite long and if there are
too many tracks in the carts, the request will timeout. When the carts
fetch fails, the users are unable to add tracks to the carts.

To fix these issues, the '/me/carts' request will return only cart
details. The UI will fetch the details for a cart when selected.