Make store__track unique constraints more restrictive;
Previously the unique constraint included the track_id, which resulted in
multipe tracks being created for the same track in the store (i.e. for multiple
store__tracks with the same store__track_store_id). There should not be a reason
to have multiple store__track rows for the same track in a store and thus
there should not be a need to include the track_id in the constraint.