Fix Beatport artist, label and playlist details fetching;
As the site structure has changed with the recent update, the name
is no longer included in the DOM for artists and labels. Instead the
name must be scraped from the _next_data JSON.

The code was already previously broken in that when "searching" an
artist, label of playlist to follow, the image shown in the results was
broken. This commit fixes that for artists and labels. For playlists
there might not be a suitable image and something else should be used
(perhaps Beatport logo?)