Implement support for debug on mobile devices;
Google (and probably Spotify) cannot be configured to allow ip addresses
for return urls. Also https is required, so the easiest solution is to
use a tunneling service like ngrok. Ngrok however does not support
redirecting different paths to different ports, so a frontend proxy
(or alternatively a backend proxy / serving of front) is required.

In order to start the frontend for debugging, add a domain in Ngrok UI
and run:

> API_URL=https://your_public_url.ngrok-free.app/api\
FRONTEND_URL=https://your_public_url.ngrok-free.app\
yarn start

> ngrok http --domain=your_public_url.ngrok-free.app 4004