Ensure follow popup buttons are updated correctly;
Previously the url was used as the key and entities without url were
rendered thus causing multiple buttons to have null as the key. The url
should only be empty for Bandcamp entities and following those without
the url is not possible. Thus it does not make sense to show the buttons
for entities without urls.