Improve Bandcamp track details extraction heuristics;
Earlier tracks with a title containing parentheses such as track two in
https://ausmusic.bandcamp.com/album/the-beauty-and-the-beat the parsing
returned an empty title. Now the heuristic should modify the title only
if the parentheses include featuring artists or version info.

This commit also improves the extraction of details from tracks that
have the version (e.g. remix) and featuring info in the title. The
expectation is that the featuring information always comes first
and the version info after that. Not sure if this is a valid assumption
but at least it should be better than it was. Having the version info
included is more important than having the featuring info as it
distinguishes the track more.