Define track number by index for Spotify albums with multiple discs;
The release__track table does not store the disc number nor do the stores
other than Spotify provide that info. Thus it is better to use the track
index in the album's tracks list to define the track number expecting
the tracks are sorted (because why would they not be?). Before this change
inserts of tracks for albums with multiple discs would fail because of
a track number constraint conflict.