Do not rollback track inserts on track number conflict;
As the track number updating is done in the same transaction as the inserts,
previously the whole batch of tracks would not get inserted if the track
number update would fail. As this is an error that will not go away with
retries, it is better to ignore the track number than prevent the insertion
of the tracks.