Fix Bandcamp follows;
The store__artist_url is not set for (some?) Bandcamp artists and thus
the store__artist_id was not returned from ensureArtistExists.

This introduces a hardcoded Bandcamp url to prevent circular dependencies
between the store logic and the database code. This should be refactored
in the future.

Also, the store__artist_url did not get updated for the Bandcamp artists
if the artist had already been added to the database without a watch and
thus the follow was not returned to the frontend.