Test how the system breaks without preview_url filter for Spotify tracks;
Spotify decided to remove the preview_url from the responses
(https://developer.spotify.com/blog/2024-11-27-changes-to-the-web-api)
and thus all Spotify processing will return 0 tracks and the
logs get filled with warnings. Although it does not seem to be
possible to get the preview_urls for the tracks from the API
it would be good to at least be able to get the tracks in the
database in order to link to the tracks and create playlists.

The preview_url is returned to the browser in the web UI.
Probably not possible to create a similar token on the
backend (and definitely a breach of the user agreement).