Skip tracks without previews in player;
Spotify tracks no longer have previews, but will still get added
to the database. It is useful to see the new tracks even if those
cannot be listened to in the player.

This commit changes the next track jump to skip all tracks that
have no previews (null preview url).