Extract cart dropdown and fix styles;
After extracting, the button styles did not get included in the build
anymore and thus an import was required. Would probably make sense to
create a separate Button component that would import the buttons.css.

The pill buttons in the popups also broke for some reason. For some
reason the button-push_button_label was used for the items and that made
the texts to not be displayed on narrow screens. Also the pill content
alignment was wrong, so align-items and justify-content were added to
the pill class.