Fix track_details function for Bandcamp tracks;
Most Bandcamp tracks in the database do not have the preview url set
as it needs to be fetched when requesting the preview (requires the
token parameter). With the preview url not null filter, the
track_details function would no longer return rows for tracks available
only on Bandcamp. It would also not include the Bandcamp previews if the
preview urls were not set.