Fix / hack the positioning of the bottom scroll to current button;
Ideally the button would be positioned with 'position: fixed', but that
will position it below the footer. 'position: fixed' will position it
correctly, but it will also reserve space from the flow. By wrapping the
button inside an absolute positioned container, the button is removed
from the flow and the space is not reserved. There however is a small
extra empty space at the bottom of the view still, probably caused by
the margins in the tracks table.